% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators-composite.R
\name{g_sampled_from}
\alias{g_sampled_from}
\title{Generate one of a fixed set of values}
\usage{
g_sampled_from(values)
}
\arguments{
\item{values}{An atomic vector with at least one element.}
}
\value{
A generator descriptor (class \code{hegelr_generator}) drawing one
element of \code{values} per case.
}
\description{
Draws an index with \code{generate_integer} inside a
SAMPLED_FROM span and returns the corresponding element of the atomic
vector \code{values}.
}
\examples{
g_sampled_from(c("alpha", "beta", "gamma"))
}
